only do tailwindcss stuff if tailwind option was selected

+++ generate.js
@@ -414,18 +414,22 @@ export class DemoGenerator {
 
       const tmpdir = os.tmpdir()
       fs.writeFileSync(`${tmpdir}/input.css`, input + '}')
-      const config = await ejs.renderFile(path.join(DemoGenerator.templates, 'tailwind.config.js'), this)
-      fs.writeFileSync(`${tmpdir}/tailwind.config.js`, config)
-      execSync(`npx tailwindcss -c ${tmpdir}/tailwind.config.js -i ${tmpdir}/input.css -o ${tmpdir}/index.css`, { stdio: 'pipe' })
-      proposed = fs.readFileSync(path.join(tmpdir, 'index.css'), 'utf-8')
+
+      if (this.options.tailwindcss) {
+        const config = await ejs.renderFile(path.join(DemoGenerator.templates, 'tailwind.config.js'), this)
+        fs.writeFileSync(`${tmpdir}/tailwind.config.js`, config)
+        execSync(`npx tailwindcss -c ${tmpdir}/tailwind.config.js -i ${tmpdir}/input.css -o ${tmpdir}/index.css`, { stdio: 'pipe' })
+        proposed = fs.readFileSync(path.join(tmpdir, 'index.css'), 'utf-8')
+
+        await this.#rmFile('tailwind.config.js')
+        await this.#rmFile('src/input.css')
+
+        fs.unlinkSync(`${tmpdir}/input.css`)
+        fs.unlinkSync(`${tmpdir}/index.css`)
+        fs.unlinkSync(`${tmpdir}/tailwind.config.js`)
+      }
+
       await this.#writeFile('public/index.css', proposed)
-
-      await this.#rmFile('tailwind.config.js')
-      await this.#rmFile('src/input.css')
-
-      fs.unlinkSync(`${tmpdir}/input.css`)
-      fs.unlinkSync(`${tmpdir}/index.css`)
-      fs.unlinkSync(`${tmpdir}/tailwind.config.js`)
 
       if (!fs.existsSync('.git')) {
         execSync('git init -b main', { stdio: 'inherit' })
